system_scripts/compression/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
snac-audio>=0.1.0
ffmpeg-python>=0.2.0
numpy>=1.24.0